+++ client/components/BusinessAnalytics/AvgRevPerGuest.js
