+++ client/store/businessAnalyticsReducer.js
@@ -3,7 +3,6 @@ import axios from 'axios'
 /**
  * ACTION TYPES
  */
-const GET_STOCK_QUERY_RESULTS = 'GET_STOCK_QUERY_RESULTS'
 const GET_NUM_ORDERS_PER_HOUR_RESULTS = 'GET_NUM_ORDERS_PER_HOUR_RESULTS'
 const GET_AVG_REVENUE_PER_GUEST = 'GET_AVG_REVENUE_PER_GUEST'
 const GET_TIP_PERCENTAGE_CHART = 'GET_TIP_PERCENTAGE_CHART'
@@ -13,7 +12,6 @@ const GET_MENU_SALES_NUMBERS_CHART = 'GET_MENU_SALES_NUMBERS_CHART'
  * INITIAL STATE
  */
 const initialState = {
-  stockQueries: {},
   numOrdersPerHour: {
     year: [],
     month: [],
@@ -37,10 +35,7 @@ const initialState = {
 /**
  * ACTION CREATORS
  */
-const gotStockQueryResults = queryResults => ({
-  type: GET_STOCK_QUERY_RESULTS,
-  queryResults
-})
+
 const gotNumOrdersPerHour = (results, timeInterval) => ({
   type: GET_NUM_ORDERS_PER_HOUR_RESULTS,
   results,
@@ -64,15 +59,6 @@ const gotMenuSalesNumbersChart = chartResults => ({
 /**
  * THUNK CREATORS
  */
-export const getStockQueryResults = () => async dispatch => {
-  try {
-    const res = await axios.get('/api/tables/stockQueries')
-    console.log(res.data)
-    dispatch(gotStockQueryResults(res.data))
-  } catch (err) {
-    console.error(err)
-  }
-}
 
 export const getNumOrdersPerHour = timeInterval => async dispatch => {
   try {
@@ -125,11 +111,6 @@ export const getMenuSalesNumbersChart = timeInterval => async dispatch => {
  */
 export default function(state = initialState, action) {
   switch (action.type) {
-    case GET_STOCK_QUERY_RESULTS:
-      return {
-        ...state,
-        stockQueries: action.queryResults
-      }
     case GET_NUM_ORDERS_PER_HOUR_RESULTS:
       return {
         ...state,
