+++ client/store/index.js
@@ -3,21 +3,15 @@ import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './userReducer'
-import waiters from './waiterReducer'
-import orders from './orderReducer'
-import menu from './menuReducer'
 import summary from './summaryReducer'
-import stockQueries from './stockQueryReducer'
-import analytics from './analyticsReducer'
+import businessAnalytics from './businessAnalyticsReducer'
+import customizedQuery from './customizedQueryReducer'
 
 const reducer = combineReducers({
   user,
   summary,
-  waiters,
-  menu,
-  orders,
-  stockQueries,
-  analytics
+  businessAnalytics,
+  customizedQuery
 })
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
