+++ server/api/index.js
@@ -14,11 +14,9 @@ client.connect()
 module.exports = router
 
 router.use('/users', require('./users'))
-router.use('/waiters', require('./waiters'))
-router.use('/orders', require('./orders'))
-router.use('/menu', require('./menu'))
-router.use('/tables', require('./tables'))
-router.use('/analytics', require('./analytics'))
+router.use('/businessAnalytics', require('./businessAnalytics'))
+router.use('/summary', require('./summary'))
+router.use('/customizedQuery', require('./customizedQuery'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
