--- server/api/waiters.js
@@ -1,49 +1,0 @@
-const router = require('express').Router()
-const pg = require('pg')
-const config = 'postgres://yourname:yourpassword@localhost:5432/nestegg'
-const client = new pg.Client(config)
-client.connect()
-
-module.exports = router
-
-router.get('/', async (req, res, next) => {
-  try {
-    const allWaiters = await client.query('SELECT * FROM WAITERS')
-    res.json(allWaiters)
-  } catch (error) {
-    next(error)
-  }
-})
-
-router.get('/fields', async (req, res, next) => {
-  try {
-    const waiterFields = await client.query(
-      'SELECT * FROM WAITERS WHERE ID = 1' // <---
-    )
-    res.json(waiterFields.fields)
-  } catch (error) {
-    next(error)
-  }
-})
-
-router.get('/query/:field/:input', async (req, res, next) => {
-  try {
-    const {field, input} = req.params
-    const queriedWaiters = await client.query(
-      `SELECT * FROM WAITERS WHERE ${field} = '${input}'` //add LIMIT
-    )
-    res.json(queriedWaiters)
-  } catch (error) {
-    next(error)
-  }
-})
-
-//SEQUELIZE GET REQUEST (FOR NOW...) GETS ALL ENUM TYPES
-const {Waiter} = require('../db/models')
-router.get('/enumValues', async (req, res, next) => {
-  try {
-    res.json(Waiter.rawAttributes.sex.values)
-  } catch (err) {
-    next(err)
-  }
-})
